pyproject.toml
[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[project]
name = "tidy"
version = "0.1.0"
description = "Tidy is a Python package designed to review your dbt project and enforce validation checks for project structure and standards."
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "pydantic (>=2.10.6,<3.0.0)"
]
license = "MIT"
license-files = ["LICENSE"]
authors = [
    {name = "Cameron Cyr"}
]
keywords = ["dbt"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

[project.urls]
Homepage = "https://github.com/JustABitOff/tidy"

[tool.setuptools.packages.find]
where = ["tidy"]
[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
ruff = "^0.9.6"
factory-boy = "^3.3.3"
coverage = "^7.6.12"

